Catch debug-mode exceptions
These changes aim to catch the exceptions that are displayed in the console when leaving Karma in debug mode. In some cases, such as the changes that affect adapters, it may be necessary to alter what happens in the catch block. For now the goal was to catch exceptions with minimal changes to the existing code. These changes do not completely eliminate, but may reduce the flaky failures referenced in #642.